修复: 解决三层架构编译错误
- 移除 RuntimeState 中无效的 serde 属性
- 修正 MemoryScope/PrivacyLevel/MemoryStatus 枚举变体
- 修复 generate_id() 调用添加 prefix 参数
- 修正 KeyEntity/RuntimeContextDigest/WriteBackRequest 结构体字段
- 修复 usize/u32 类型不匹配
- 修复 deny_request 借用检查器错误
- 补充 MemorySummary.semantic_context 和 MemoryQuery 缺失字段
- 清理未使用的导入

+++ src-tauri/src/shell_agent/state.rs
@@ -20,10 +20,10 @@
 //! ```
 
 use std::path::{Path, PathBuf};
-use std::sync::{Arc, Mutex};
+use std::sync::Mutex;
 
 use crate::memory::MemoryService;
-use crate::rule_engine::{RuleEngine, RuleGenerator, Rule};
+use crate::rule_engine::{RuleEngine, RuleGenerator};
 use crate::permission::{PermissionChecker, PermissionStore, PermissionState, PermissionScope, GrantSource};
 
 /// 三层架构统一状态
@@ -117,9 +117,9 @@ impl BehaviorState {
 
         // 1. 运行记忆维护
         let memory_result = self.memory_service.run_maintenance();
-        result.memory_decayed = memory_result.decayed;
-        result.memory_merged = memory_result.merged;
-        result.memory_pruned = memory_result.pruned;
+        result.memory_decayed = memory_result.decayed as usize;
+        result.memory_merged = memory_result.merged as usize;
+        result.memory_pruned = memory_result.pruned as usize;
 
         // 2. 从记忆生成规则
         if let Ok(count) = self.generate_rules_from_memories() {
